Initial take on design improvements
