+++ projects/plugins/protect/src/js/components/admin-page/index.jsx
@@ -1,12 +1,18 @@
-import { AdminPage as JetpackAdminPage, Container } from '@automattic/jetpack-components';
+import {
+	AdminPage as JetpackAdminPage,
+	Container,
+	useBreakpointMatch,
+} from '@automattic/jetpack-components';
 import { useConnection } from '@automattic/jetpack-connection';
 import { __ } from '@wordpress/i18n';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import useNotices from '../../hooks/use-notices';
 import useWafData from '../../hooks/use-waf-data';
 import Logo from '../logo';
 import Notice from '../notice';
+import OnboardingPopover from '../onboarding-popover';
+import ScanButton from '../scan-button';
 import Tabs, { Tab } from '../tabs';
 import styles from './styles.module.scss';
 
@@ -15,6 +21,10 @@ const AdminPage = ( { children } ) => {
 	const { isRegistered } = useConnection();
 	const { isSeen: wafSeen } = useWafData();
 	const navigate = useNavigate();
+	const [ isSm ] = useBreakpointMatch( 'sm' );
+
+	const [ dailyAndManualScansPopoverAnchor, setDailyAndManualScansPopoverAnchor ] =
+		useState( null );
 
 	// Redirect to the setup page if the site is not registered.
 	useEffect( () => {
@@ -28,7 +38,22 @@ const AdminPage = ( { children } ) => {
 	}
 
 	return (
-		<JetpackAdminPage moduleName={ __( 'Jetpack Protect', 'jetpack-protect' ) } header={ <Logo /> }>
+		<JetpackAdminPage
+			moduleName={ __( 'Jetpack Protect', 'jetpack-protect' ) }
+			header={
+				<div className={ styles.header }>
+					<Logo />
+					<div>
+						<ScanButton ref={ setDailyAndManualScansPopoverAnchor } />
+						<OnboardingPopover
+							id="paid-daily-and-manual-scans"
+							position={ isSm ? 'bottom left' : 'middle left' }
+							anchor={ dailyAndManualScansPopoverAnchor }
+						/>
+					</div>
+				</div>
+			}
+		>
 			{ notice && <Notice floating={ true } dismissable={ true } { ...notice } /> }
 			<Container horizontalSpacing={ 0 }>
 				<Tabs className={ styles.navigation }>
