+++ projects/plugins/protect/src/js/components/admin-page/styles.module.scss
@@ -14,3 +14,8 @@
     position: relative;
     top: -1px;
 }
+
+.header {
+    display: flex;
+    justify-content: space-between;
+}
