+++ projects/plugins/protect/src/js/components/scan-header/styles.module.scss
@@ -1,0 +1,132 @@
+svg.spinner {
+    width: 32px;
+    height: 32px;
+    margin: 0;
+    color: var( --jp-black );
+}
+
+.popover-text {
+	width: 250px;
+	padding: calc( var( --spacing-base ) * 2 ); // 16px
+}
+
+.loading-text {
+    font-size: 18px;
+}
+
+.firewall-header {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.firewall-subheading {
+    display: flex;
+    justify-content: flex-start;
+    flex-wrap: wrap;
+    align-items: center;
+
+    .brute-force-protection-subheading {
+        margin-right: calc( var( --spacing-base ) / 2 ); // 4px
+    }
+
+    &__content {
+        display: flex;
+    }
+}
+
+.upgrade-button {
+    margin-top: calc( var( --spacing-base ) * 4 ); // 32px
+}
+
+.stat-card-wrapper {
+    display: flex;
+    margin-left: auto;
+    flex-wrap: wrap;
+
+    >:first-child {
+        margin-right: calc( var( --spacing-base ) * 3 ); // 24px
+    }
+
+    .disabled {
+        opacity: 0.5;
+    }
+}
+
+.stat-card-icon {
+    width: 100%;
+    margin-bottom: calc( var( --spacing-base ) * 3 ); // 24px
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    .shield {
+        fill: var( --jp-red-40 ) ;
+    }
+}
+
+.stat-card-label {
+    white-space: nowrap;
+}
+
+@media ( max-width: 1115px ) {
+    .firewall-header {
+        display: inline-block;
+    }
+
+    .stat-card-wrapper {
+        margin-top: calc( var( --spacing-base ) * 3 ); // 24px
+    }
+}
+
+@media ( max-width: 599px ) {
+    .firewall-heading {
+        width: 90%;
+    }
+    
+    .stat-card-wrapper {
+        margin-top: calc( var( --spacing-base ) * 3 ); // 24px
+
+        >:first-child {
+            margin-right: 0;
+            margin-bottom: var( --spacing-base ); // 8px
+        }
+    }
+
+    .stat-card-icon {
+        margin-bottom: 0;
+    }
+}
+
+.scan-section-header {
+	> :first-child {
+		flex: 1;
+	}
+
+	@media ( min-width: 960px ) {
+		display: flex;
+		align-items: flex-start;
+	}
+}
+
+.scan-section-header__content {
+	display: flex;
+	flex-direction: column;
+	gap: var( --spacing-base ); // 8px
+}
+
+.scan-section-header__subtitle {
+	display: flex;
+	align-items: center;
+	margin-bottom: 0;
+}
+
+.scan-section-header__icon {
+	margin-left: var( --spacing-base ); // 8px
+}
+
+.scan-section-header__title {
+	display: flex;
+	align-items: center;
+}
