+++ projects/plugins/protect/src/js/components/summary/index.jsx
@@ -1,24 +1,15 @@
-import { useBreakpointMatch } from '@automattic/jetpack-components';
 import { dateI18n } from '@wordpress/date';
 import { __, sprintf } from '@wordpress/i18n';
-import React, { useState } from 'react';
-import usePlan from '../../hooks/use-plan';
 import useProtectData from '../../hooks/use-protect-data';
-import ScanSectionHeader from '../../routes/scan/scan-section-header';
-import OnboardingPopover from '../onboarding-popover';
+import ScanSectionHeader from '../scan-header/scan-section-header';
 
 const Summary = () => {
-	const [ isSm ] = useBreakpointMatch( 'sm' );
 	const {
 		counts: {
 			current: { threats: numThreats },
 		},
 		lastChecked,
 	} = useProtectData();
-	const { hasPlan } = usePlan();
-
-	// Popover anchors
-	const [ dailyScansPopoverAnchor, setDailyScansPopoverAnchor ] = useState( null );
 
 	return (
 		<ScanSectionHeader
@@ -32,24 +23,11 @@ const Summary = () => {
 					  )
 					: undefined
 			}
-			subtitle={
-				<>
-					<div ref={ setDailyScansPopoverAnchor }>
-						{ sprintf(
-							/* translators: %s: Latest check date  */
-							__( 'Latest results as of %s', 'jetpack-protect' ),
-							dateI18n( 'F jS', lastChecked )
-						) }
-					</div>
-					{ ! hasPlan && (
-						<OnboardingPopover
-							id="free-daily-scans"
-							position={ isSm ? 'bottom' : 'middle right' }
-							anchor={ dailyScansPopoverAnchor }
-						/>
-					) }
-				</>
-			}
+			subtitle={ sprintf(
+				/* translators: %s: Latest check date  */
+				__( 'Latest results as of %s', 'jetpack-protect' ),
+				dateI18n( 'F jS', lastChecked )
+			) }
 		/>
 	);
 };
