+++ projects/plugins/protect/src/js/routes/scan/index.jsx
@@ -1,14 +1,9 @@
-import { AdminSectionHero, Container, Col, H3, Text } from '@automattic/jetpack-components';
-import { useConnectionErrorNotice, ConnectionError } from '@automattic/jetpack-connection';
-import { Spinner } from '@wordpress/components';
-import { __, sprintf } from '@wordpress/i18n';
-import { useMemo } from 'react';
-import inProgressImage from '../../../../assets/images/in-progress.png';
+import { AdminSection, Container, Col, useBreakpointMatch } from '@automattic/jetpack-components';
+import { useState } from 'react';
 import AdminPage from '../../components/admin-page';
-import ErrorScreen from '../../components/error-section';
-import ProgressBar from '../../components/progress-bar';
+import OnboardingPopover from '../../components/onboarding-popover';
 import ScanFooter from '../../components/scan-footer';
-import SeventyFiveLayout from '../../components/seventy-five-layout';
+import ScanHeader from '../../components/scan-header';
 import Summary from '../../components/summary';
 import ThreatsList from '../../components/threats-list';
 import useScanStatusQuery, { isScanInProgress } from '../../data/scan/use-scan-status-query';
@@ -16,142 +11,13 @@ import useAnalyticsTracks from '../../hooks/use-analytics-tracks';
 import { OnboardingContext } from '../../hooks/use-onboarding';
 import usePlan from '../../hooks/use-plan';
 import useProtectData from '../../hooks/use-protect-data';
-import useWafData from '../../hooks/use-waf-data';
 import onboardingSteps from './onboarding-steps';
-import ScanSectionHeader from './scan-section-header';
-import styles from './styles.module.scss';
-
-const ConnectionErrorCol = () => {
-	const { hasConnectionError } = useConnectionErrorNotice();
-
-	return (
-		<>
-			{ hasConnectionError && (
-				<Col className={ styles[ 'connection-error-col' ] }>
-					<ConnectionError />
-				</Col>
-			) }
-			<Col>
-				<div id="jp-admin-notices" className="my-jetpack-jitm-card" />
-			</Col>
-		</>
-	);
-};
-
-const HeaderContainer = () => {
-	return (
-		<Container horizontalSpacing={ 0 }>
-			<ConnectionErrorCol />
-		</Container>
-	);
-};
-
-const ErrorSection = ( { errorMessage, errorCode } ) => {
-	return (
-		<>
-			<HeaderContainer />
-			<Container horizontalSpacing={ 3 } horizontalGap={ 4 }>
-				<Col>
-					<ScanSectionHeader />
-				</Col>
-				<Col>
-					<ErrorScreen
-						baseErrorMessage={ __(
-							'We are having problems scanning your site.',
-							'jetpack-protect'
-						) }
-						errorMessage={ errorMessage }
-						errorCode={ errorCode }
-					/>
-				</Col>
-			</Container>
-		</>
-	);
-};
-
-const ScanningSection = ( { currentProgress } ) => {
-	const { hasPlan } = usePlan();
-	const { globalStats } = useWafData();
-	const totalVulnerabilities = parseInt( globalStats?.totalVulnerabilities );
-	const totalVulnerabilitiesFormatted = isNaN( totalVulnerabilities )
-		? '50,000'
-		: totalVulnerabilities.toLocaleString();
-
-	return (
-		<>
-			<HeaderContainer />
-			<Container horizontalSpacing={ 3 } horizontalGap={ 4 }>
-				<Col>
-					<ScanSectionHeader />
-				</Col>
-				<Col>
-					<SeventyFiveLayout
-						main={
-							<div className={ styles[ 'main-content' ] }>
-								<Container horizontalSpacing={ 0 } horizontalGap={ 7 } fluid={ true }>
-									<Col className={ styles[ 'loading-content' ] }>
-										<Spinner
-											style={ {
-												color: 'black',
-												marginTop: 0,
-												marginLeft: 0,
-											} }
-										/>
-										<span>{ __( 'Scanning your site…', 'jetpack-protect' ) }</span>
-									</Col>
-									<Col>
-										<H3 style={ { textWrap: 'balance' } }>
-											{ __( 'Your results will be ready soon', 'jetpack-protect' ) }
-										</H3>
-										{ hasPlan && <ProgressBar value={ currentProgress || 0 } /> }
-										<Text>
-											{ sprintf(
-												// translators: placeholder is the number of total vulnerabilities i.e. "22,000".
-												__(
-													'We are scanning for security threats from our more than %s listed vulnerabilities, powered by WPScan. This could take a minute or two.',
-													'jetpack-protect'
-												),
-												totalVulnerabilitiesFormatted
-											) }
-										</Text>
-									</Col>
-								</Container>
-							</div>
-						}
-						secondary={
-							<div className={ styles.illustration }>
-								<img src={ inProgressImage } alt="" />
-							</div>
-						}
-						preserveSecondaryOnMobile={ false }
-						fluid={ true }
-					/>
-				</Col>
-			</Container>
-		</>
-	);
-};
-
-const DefaultSection = () => {
-	return (
-		<>
-			<HeaderContainer />
-			<Container horizontalSpacing={ 3 } horizontalGap={ 4 }>
-				<Col>
-					<Summary />
-				</Col>
-				<Col>
-					<ThreatsList />
-				</Col>
-			</Container>
-		</>
-	);
-};
 
 const ScanPage = () => {
 	const { hasPlan } = usePlan();
 	const { lastChecked } = useProtectData();
 	const { data: status } = useScanStatusQuery( { usePolling: true } );
+	const [ isSm ] = useBreakpointMatch( 'sm' );
 
 	let currentScanStatus;
 	if ( status.error ) {
@@ -171,22 +37,40 @@ const ScanPage = () => {
 		},
 	} );
 
-	const renderSection = useMemo( () => {
-		if ( isScanInProgress( status ) ) {
-			return <ScanningSection currentProgress={ status.currentProgress || 0 } />;
-		}
-
-		if ( status.error ) {
-			return <ErrorSection errorMessage={ status.errorMessage } errorCode={ status.errorCode } />;
-		}
-
-		return <DefaultSection />;
-	}, [ status ] );
+	// Popover anchors
+	const [ dailyScansPopoverAnchor, setDailyScansPopoverAnchor ] = useState( null );
 
 	return (
 		<OnboardingContext.Provider value={ onboardingSteps }>
 			<AdminPage>
-				<AdminSectionHero>{ renderSection }</AdminSectionHero>
+				<ScanHeader
+					isScanning={ isScanInProgress( status ) }
+					currentProgress={ status.currentProgress }
+					error={ status.error }
+					errorMessage={ status.errorMessage }
+					errorCode={ status.errorCode }
+					summary={
+						<>
+							<div ref={ setDailyScansPopoverAnchor }>
+								<Summary />
+							</div>
+							{ ! hasPlan && (
+								<OnboardingPopover
+									id="free-daily-scans"
+									position={ isSm ? 'bottom' : 'middle right' }
+									anchor={ dailyScansPopoverAnchor }
+								/>
+							) }
+						</>
+					}
+				/>
+				<AdminSection>
+					<Container horizontalSpacing={ 7 } horizontalGap={ 3 }>
+						<Col>
+							<ThreatsList />
+						</Col>
+					</Container>
+				</AdminSection>
 				<ScanFooter />
 			</AdminPage>
 		</OnboardingContext.Provider>
