+++ projects/plugins/protect/src/js/routes/scan/styles.module.scss
@@ -28,48 +28,3 @@
 	justify-content: flex-end;
 	margin-top: calc( var( --spacing-base ) * 3 + 1px ); // 25px
 }
-
-.scan-section-header {
-	> :first-child {
-		flex: 1;
-	}
-
-	@media ( min-width: 960px ) {
-		display: flex;
-		align-items: flex-start;
-	}
-}
-
-.scan-section-header__content {
-	display: flex;
-	flex-direction: column;
-	gap: var( --spacing-base ); // 8px
-
-	> :last-child {
-		margin-top: calc( var( --spacing-base ) * 2 ); // 16px
-	}
-}
-
-.scan-section-header__subtitle {
-	display: flex;
-	align-items: center;
-	color: var( --jp-black );
-	margin-bottom: 0;
-}
-
-.scan-section-header__icon {
-	margin-left: -4px; // this fix a blank space on right/left from @wordpress/icons
-	margin-right: var( --spacing-base ); // 8px
-}
-
-.scan-section-header__controls {
-	display: flex;
-	flex-direction: column;
-	align-items: flex-end;
-	gap: calc( var( --spacing-base ) * 2 ); // 16px
-}
-
-.scan-section-header__controls__row {
-	display: flex;
-	gap: calc( var( --spacing-base ) * 2 ); // 16px
-}
